pyproject.toml
[project]
name = "fordefi-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ecdsa>=0.19.0",
    "fastapi>=0.115.5",
    "pydantic>=2.10.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "uvicorn>=0.32.1",
]
